Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
onnxruntime==1.23.1
numpy==1.24.3
Pillow==10.0.1
python-multipart==0.0.6
